Retrieve Tides Data From NOAA CO-OPS API

Retrieve Tides Data From NOAA CO-OPS API

query_coops_data(
  station_id,
  start_date,
  end_date,
  data_product,
  units = "english",
  time_zone = "gmt",
  datum = NULL,
  interval = NULL,
  bin = NULL
)

Arguments

station_id
is a character string that provides the a 7 character station id.

start_date
is a character string that specifies the start date for the retrieval period. Dates can be specified in the following formats: yyyyMMdd, yyyyMMdd HH:mm, MM/dd/yyyy, or MM/dd/yyyy HH:mm.

end_date
is a character string that specifies the end date for the retrieval period. Dates can be specified in the following formats: yyyyMMdd, yyyyMMdd HH:mm, MM/dd/yyyy, or MM/dd/yyyy HH:mm.

data_product
specifies the data product to be returned. See CO-OPS API Documentation for the available data products.

units
a character string specifying if the data should be returned using metric or English units. Defaults to 'english'.

time_zone
a character string specifying what time zone information the data should be returned with. Options include Greenwich Mean Time 'gmt', Local Standard Time 'lst', and Local Standard/Local Daylight Time 'lst_ldt'. Local times refer to the local time of the specified station. The default is 'gmt'

datum
a character string indicating the datum that should be returned. See CO-OPS API Documentation for the available datums.

interval
a character string that specifies the interval for which Meteorological data is returned. The API defaults to every six minutes and does not need to be specified. Other option include hourly 'h' and 'hilo'. The retrieval time period specified by start_date and end_date to create restrictions on the intervals that can be returned. See CO-OPS API Documentation for details

bin
the bin number for the indicated currents station. If a bin is not specified for a PORTS station, the data is returned using a predefined real-time bin.

Value

a data frame.

Examples

# \donttest{
# Do Not Run
a <- query_coops_data('9414290',
                      '20170101',
                      '20170201',
                      'predictions',
                       interval = 'hilo',
                       datum = 'MLLW')
# }