Find All NOAA Stations

This function produces a data frame with all NOAA stations. The list of stations is retrieved from NOAA's website when the function is called.

list_coops_stations()

Value

A data frame.

Details

In the returned data frame there is one row for each station. The name, location and date that the station was established are included as columns. In addition, there are columns that provide the status of various sensors at the station is included. The column names indicate the type of sensor

In the status columns a value of 1 indicates that sensor is working A 0 indicates that the sensor is not working. If a particular station does not have the capability indicated by the column name, the value provided is NA

Examples

# Do Not Run
# \donttest{
station_df <- list_coops_stations()
# }