Query Data From NOAA's CO-OPS Derived Product API

Provides access to data available from NOAA's CO-OPS Derived Product API. Four derived data products are available through the API: 1) Top Ten Water Levels, 2) Annual Flood Days, 3) Extreme Water Levels and 4) Sea Level Trends. More detail about each data product is available with the API's documentation

query_derived_products(
  station_id = NULL,
  product_name = NULL,
  year = NULL,
  affil = NULL,
  units = "english"
)

Arguments

station_id
an optional string that provides the a 7 character station id. If omitted the derived product API returns data for all stations.

product_name
a string providing the name of the derived data product. Derived products include Top Ten Water Levels ('toptenwaterlevels'), Annual Flood Days ('annualflooddays'), Extreme Water Levels ('extremewaterlevels') and Sea Level Trends ('sealeveltrends').

year
an optional string used to limit the results from the annual flood days product to the indicated year. The argument is ignored if used with other data products.

affil
an optional argument used to limit the results from the Sea Level Trends product to U.S. ('US') or Global stations ('Global'). If omitted with Sea Level Trends all stations are returned. The argument is ignored if used with other data products.

units
a character string specifying if the data should be returned using metric or English units. Defaults to 'english'

Value

a data frame. See the documentation for the specific product for the details an it's content.

Details

Results are returned as a data frame specific to each product. For the variables returned with each endpoint see the appropriate section in the documentation.

Top Ten Water Levels

Annual Flood Days

Extreme Water Levels

Sea Level Trends

Examples

# \donttest{
# Not Run
query_derived_products(station_id = "1611400",
                       product_name = "sealeveltrends")

#>   stationId    stationName affil latitude longitude trend trendError units
#> 1   1611400 Nawiliwili, HI    US  21.9544 -159.3561  1.78       0.21 mm/yr
#>    startDate    endDate
#> 1 01/15/1955 12/15/2020

# }